+++ apps/web-stride/src/screens/home/components/blocks/types.ts
@@ -1,0 +1,15 @@
+export interface BlockType {
+  height: number;
+  txs: number;
+  timestamp: string;
+  proposer: string;
+  consumer: string;
+  hash: string;
+}
+
+export interface BlocksState {
+  loading: boolean;
+  items: BlockType[];
+}
+
+export type ItemType = BlockType;